faculty-launcher/server/requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
redis==5.0.3
minio==7.2.5
pydantic-settings==2.2.1
python-multipart==0.0.9
slowapi==0.1.9
pyjwt==2.8.0
alembic==1.13.1
rarfile==4.1